Complete sync and async fs access and exists tests
