+++ src/node/fs.ts
@@ -4,14 +4,108 @@
 
 import * as promises from 'node-internal:internal_fs_promises';
 import * as constants from 'node-internal:internal_fs_constants';
-import * as syncMethods from 'node-internal:internal_fs_sync';
 import * as callbackMethods from 'node-internal:internal_fs_callback';
 import { WriteStream, ReadStream } from 'node-internal:internal_fs_streams';
 import { Dirent, Dir } from 'node-internal:internal_fs';
 
 export * from 'node-internal:internal_fs_callback';
-export * from 'node-internal:internal_fs_sync';
-export { constants, promises, Dirent, Dir, WriteStream, ReadStream };
+import {
+  accessSync,
+  existsSync,
+  appendFileSync,
+  chmodSync,
+  chownSync,
+  closeSync,
+  copyFileSync,
+  cpSync,
+  fchmodSync,
+  fchownSync,
+  fdatasyncSync,
+  fstatSync,
+  fsyncSync,
+  ftruncateSync,
+  futimesSync,
+  //globSync
+  lchmodSync,
+  lchownSync,
+  lutimesSync,
+  linkSync,
+  lstatSync,
+  mkdirSync,
+  mkdtempSync,
+  opendirSync,
+  openSync,
+  readdirSync,
+  readFileSync,
+  readlinkSync,
+  readSync,
+  readvSync,
+  realpathSync,
+  renameSync,
+  rmdirSync,
+  rmSync,
+  statSync,
+  statfsSync,
+  symlinkSync,
+  truncateSync,
+  unlinkSync,
+  utimesSync,
+  writeFileSync,
+  writeSync,
+  writevSync,
+} from 'node-internal:internal_fs_sync';
+
+export {
+  constants,
+  promises,
+  Dirent,
+  Dir,
+  WriteStream,
+  ReadStream,
+  accessSync,
+  existsSync,
+  appendFileSync,
+  chmodSync,
+  chownSync,
+  closeSync,
+  copyFileSync,
+  cpSync,
+  fchmodSync,
+  fchownSync,
+  fdatasyncSync,
+  fstatSync,
+  fsyncSync,
+  ftruncateSync,
+  futimesSync,
+  //globSync
+  lchmodSync,
+  lchownSync,
+  lutimesSync,
+  linkSync,
+  lstatSync,
+  mkdirSync,
+  mkdtempSync,
+  opendirSync,
+  openSync,
+  readdirSync,
+  readFileSync,
+  readlinkSync,
+  readSync,
+  readvSync,
+  realpathSync,
+  renameSync,
+  rmdirSync,
+  rmSync,
+  statSync,
+  statfsSync,
+  symlinkSync,
+  truncateSync,
+  unlinkSync,
+  utimesSync,
+  writeFileSync,
+  writeSync,
+  writevSync,
+};
 
 export default {
   constants,
@@ -20,6 +114,48 @@ export default {
   Dir,
   WriteStream,
   ReadStream,
-  ...syncMethods,
   ...callbackMethods,
+  accessSync,
+  existsSync,
+  appendFileSync,
+  chmodSync,
+  chownSync,
+  closeSync,
+  copyFileSync,
+  cpSync,
+  fchmodSync,
+  fchownSync,
+  fdatasyncSync,
+  fstatSync,
+  fsyncSync,
+  ftruncateSync,
+  futimesSync,
+  //globSync
+  lchmodSync,
+  lchownSync,
+  lutimesSync,
+  linkSync,
+  lstatSync,
+  mkdirSync,
+  mkdtempSync,
+  opendirSync,
+  openSync,
+  readdirSync,
+  readFileSync,
+  readlinkSync,
+  readSync,
+  readvSync,
+  realpathSync,
+  renameSync,
+  rmdirSync,
+  rmSync,
+  statSync,
+  statfsSync,
+  symlinkSync,
+  truncateSync,
+  unlinkSync,
+  utimesSync,
+  writeFileSync,
+  writeSync,
+  writevSync,
 };
