+++ src/workerd/api/node/tests/fs-test.wd-test
@@ -1,0 +1,15 @@
+using Workerd = import "/workerd/workerd.capnp";
+
+const unitTests :Workerd.Config = (
+  services = [
+    ( name = "fs-test",
+      worker = (
+        modules = [
+          (name = "worker", esModule = embed "fs-test.js")
+        ],
+        compatibilityDate = "2025-05-01",
+        compatibilityFlags = ["nodejs_compat", "experimental"]
+      )
+    ),
+  ],
+);
